+++ packages/vue-dot/src/rules/notBeforeDate/tests/notBeforeDate.spec.ts
@@ -1,21 +1,21 @@
-import { notAfterDate } from '../';
+import { notBeforeDate } from '../';
 
 import dayjs from 'dayjs';
 import { formatDate } from '../../../functions/formatDate';
 
-describe('notAfterDate', () => {
+describe('notBeforeDate', () => {
 	const currentDate = formatDate(dayjs());
 	const pastDate = formatDate(dayjs().subtract(1, 'year'));
 	const futureDate = formatDate(dayjs().add(1, 'year'));
 
-	const rule = notAfterDate(currentDate);
+	const rule = notBeforeDate(currentDate);
 
-	it('returns true with a past date', () => {
-		expect(rule(pastDate)).toBe(true);
+	it('returns true with a future date', () => {
+		expect(rule(futureDate)).toBe(true);
 	});
 
-	it('returns an error with a future date', () => {
-		expect(typeof rule(futureDate)).toBe('string');
+	it('returns an error with a past date', () => {
+		expect(typeof rule(pastDate)).toBe('string');
 	});
 
 	it('returns true if the value is falsy', () => {